After I couldn't focus the entire day, managed to fix the prediction function for diffwave. TODO: for res1d

+++ models/mel_spec_converters.py
@@ -40,7 +40,7 @@ class SimpleMelSpecConverter:
 
     def convert(self, slice: torch.Tensor) -> torch.Tensor:
         self.mel_spec = self.mel_spec.to(slice.device)
-        output = self.mel_spec(slice)
+        output = self.mel_spec(slice.float())
         return output
